+++ trading-common/src/main/kotlin/ru/efreet/trading/exchange/Ticker.kt
@@ -3,4 +3,4 @@ package ru.efreet.trading.exchange
 /**
  * Created by fluder on 09/02/2018.
  */
-data class Ticker(var instrument: Instrument, var highestBid:Double, var lowestAsk:Double)
+data class Ticker(var instrument: Instrument, var highestBid:Float, var lowestAsk:Float)
