+++ trading-common/src/main/kotlin/ru/efreet/trading/exchange/impl/cache/CachedExchange.kt
@@ -8,7 +8,7 @@ import java.time.ZonedDateTime
 /**
  * Created by fluder on 11/02/2018.
  */
-class CachedExchange(_name: String, _fee: Double, interval: BarInterval, val cache: BarsCache) : FakeExchange(_name, _fee, interval) {
+class CachedExchange(_name: String, _fee: Float, interval: BarInterval, val cache: BarsCache) : FakeExchange(_name, _fee, interval) {
 
 
     override fun loadBars(instrument: Instrument, interval: BarInterval, startTime: ZonedDateTime, endTime: ZonedDateTime): List<XBar> {
